flowEdit file segregated
